% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glvmlss-package.R
\docType{package}
\name{glvmlss-package}
\alias{glvmlss-package}
\alias{_PACKAGE}
\title{glvmlss: Generalized Latent Variable Models for Location, Scale, and Shape parameters}
\description{
Functions for fitting Generalized Latent Variable Models for Location, Scale, and Shape Parameters (GLVM-LSS).
}
\author{
\strong{Maintainer}: Camilo Cárdenas-Hurtado \email{c.a.cardenas-hurtado@lse.ac.uk} (\href{https://orcid.org/0009-0003-1073-0912}{ORCID})

}
\keyword{internal}
