% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glvmlss.R
\name{glvmlss}
\alias{glvmlss}
\title{Fit a Generalized latent variable model for location, scale, and shape parameters}
\usage{
glvmlss(
  data,
  family,
  mu.eq = ~Z1,
  sg.eq = NULL,
  nu.eq = NULL,
  ta.eq = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{data}{Matrix \code{(n x p)} of observed variables. Missing data should be coded as \code{NA}.}

\item{family}{String vector of size \code{p} with family for each observed variable.}

\item{mu.eq}{Formula for location parameter e.g.,  \code{mu.eq = ~ Z1}.}

\item{sg.eq}{(optional) Formula for scale parameter.}

\item{nu.eq}{(optional) Formula for shape parameter.}

\item{ta.eq}{(optional) Formula for shape parameter.}

\item{control}{List of control parameters (see 'Details').}

\item{...}{Further arguments to be passed to \code{control}.}
}
\value{
A list with components:
\itemize{
\item \code{b}: An array with the estimated intercept and slopes.
\item \code{R}: Estimated factor correlation matrix (if \code{control$corLV = T}).
\item \code{ll}: Marginal log-likelihood evaluated at \code{b} and \code{Rz}.
\item \code{seb}: Estimated standard errors for \code{b}.
\item \code{seR}: Estimated standard errors for \code{R} (if \code{control$corLV = T}).
}
}
\description{
Fit a Generalized latent variable model for location, scale, and shape parameters
}
\details{
Test
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
