% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glvmlss.R
\name{glvmlss_sim}
\alias{glvmlss_sim}
\title{Simulate data for a Generalized latent variable model for location, scale, and shape parameters}
\usage{
glvmlss_sim(
  n,
  family,
  mu.eq = ~Z1,
  sg.eq = NULL,
  nu.eq = NULL,
  ta.eq = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{n}{Number of simulated entries.}

\item{family}{String vector of size \code{p} with family for each observed variable.}

\item{mu.eq}{Formula for location parameter e.g.,  \code{mu.eq = ~ Z1}.}

\item{sg.eq}{(optional) Formula for scale parameter.}

\item{nu.eq}{(optional) Formula for shape parameter.}

\item{ta.eq}{(optional) Formula for shape parameter.}

\item{control}{List of control parameters (see 'Details').}

\item{...}{Further arguments to be passed to \code{control}.}
}
\value{
A list with components:
\itemize{
\item \code{Y}: A matrix \code{(n x p)} of observed variables, each column from a distribution specified in \code{family}.
\item \code{Z}: Simulated latent variables (includes intercept).
\item \code{b}: Original coefficients.
\item \code{R}: Original factor correlation matrix.
}
}
\description{
Simulate data for a Generalized latent variable model for location, scale, and shape parameters
}
\details{
Test
}
\author{
Camilo Cárdenas-Hurtado (\email{c.a.cardenas-hurtado@lse.ac.uk}).
}
