% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISA2018.R
\docType{data}
\name{PISA2018}
\alias{PISA2018}
\title{PISA 2018 computer-based mathematics exam data}
\format{
\subsection{\code{PISA2018}}{

A data frame with 1,280 rows and 18 columns:
\describe{
\item{T1-T9}{Response times (in log-minutes) for each item.}
\item{Y1-Y9}{Responses for items 1 to 9.}
}
}
}
\source{
OECD, 2020. Data available from \url{https://www.oecd.org/en/data/datasets/pisa-2018-database.html}.
}
\usage{
PISA2018
}
\description{
Sample of data from the 2018 PISA computer-based mathematics exam, first testing cluster.
The data set contains responses to 9 items and their corresponding response times (in log-minutes).
We only included a sample of students from Brazil who had complete answers.
}
\keyword{datasets}
